DOC : responsive-ui #45 move all types into types folder
흩어져있던 type 선언들을 모두 types 폴더로 이동함
