+++ src/api/auth.ts
@@ -1,14 +1,5 @@
 import API from './base';
-import type { User } from '../types';
-
-type loginForm = {
-    username: string;
-    password: string;
-};
-
-type signUpForm = loginForm & {
-    nickname: string;
-};
+import type { User, loginForm, signUpForm } from '../types';
 
 export const sign_up = (signUpData: signUpForm) =>
     API.post<User>(`/join`, signUpData);
@@ -16,8 +7,7 @@ export const sign_up = (signUpData: signUpForm) =>
 export const login = (loginData: loginForm) =>
     API.post<User>(`/login`, loginData);
 
-export const getMyInfo = () =>
-    API.get<User>(`/api/member/mypage`);
+export const getMyInfo = () => API.get<User>(`/api/member/mypage`);
 
 export const findByUsername = (userName: string) =>
     API.get<{ members: User[] }>(
