+++ src/api/schedule.ts
@@ -1,10 +1,5 @@
 import API from './base';
-import type { Event, UserEvent } from '../types';
-
-type GetMembersScheduleResponse = {
-    username: string;
-    schedule: UserEvent[];
-}[];
+import type { Event, UserEvent, GetMembersScheduleResponse } from '../types';
 
 export const getMySchedule = () =>
     API.get<{ schedule: UserEvent[] }>(`/api/schedule/myschedule`);
@@ -16,10 +11,14 @@ export const deleteMySchedule = (eventId: string) =>
     API.delete(`/api/schedule/myschedule?scheduleId=${eventId}`);
 
 export const getMembersSchedule = (teamId: number) =>
-    API.get<GetMembersScheduleResponse>(`/api/schedule/myteam?teamId=${teamId}`);
+    API.get<GetMembersScheduleResponse>(
+        `/api/schedule/myteam?teamId=${teamId}`,
+    );
 
 export const getTeamSchedule = (teamId: number) =>
-    API.get<{ schedule: Event[] }>(`/api/schedule/teamschedule?teamId=${teamId}`);
+    API.get<{ schedule: Event[] }>(
+        `/api/schedule/teamschedule?teamId=${teamId}`,
+    );
 
 export const createTeamSchedule = (teamId: number, newEvent: Event) =>
     API.post(`/api/schedule/teamschedule?teamId=${teamId}`, newEvent);
