+++ src/components/schedule/Schedule.tsx
@@ -1,33 +1,9 @@
 import FullCalendar from '@fullcalendar/react';
 import timeGridPlugin from '@fullcalendar/timegrid';
 import interactionPlugin from '@fullcalendar/interaction';
-import {
-    EventApi,
-    DateSelectArg,
-    EventClickArg,
-    EventHoveringArg,
-    EventAddArg,
-    EventChangeArg,
-    EventRemoveArg,
-} from '@fullcalendar/core';
 import { renderEventContent } from '../../utils';
 import { START_DATE, START_TIME } from '../../constant';
-import type { UserEvent, EventSource } from '../../types';
-// import '../../styles/componentStyle/schedule.css';
-
-type ScheduleProps = {
-    selectable?: boolean;
-    events?: UserEvent[] | undefined;
-    eventSources?: EventSource[] | undefined;
-    handleEvents?: (events: EventApi[]) => void;
-    handleDateSelect?: (selectInfo: DateSelectArg) => void;
-    handleEventClick?: (clickInfo: EventClickArg) => void;
-    handleMouseEnter?: (enterInfo: EventHoveringArg) => void;
-    handleMouseLeave?: (leaveInfo: EventHoveringArg) => void;
-    handleEventAdd?: (addInfo: EventAddArg) => void;
-    handleEventChange?: (changeInfo: EventChangeArg) => void;
-    handleEventRemove?: (removeInfo: EventRemoveArg) => void;
-};
+import type { ScheduleProps } from '../../types';
 
 const Schedule = ({
     selectable,
