+++ src/types/member.ts
@@ -1,5 +1,14 @@
-export type User = {
+type User = {
     memberId: number;
     username: string;
     nickname: string;
 };
+type loginForm = {
+    username: string;
+    password: string;
+};
+type signUpForm = loginForm & {
+    nickname: string;
+};
+
+export type { User, loginForm, signUpForm };
